+++ monkey/monkey_island/cc/resources/__init__.py
@@ -6,3 +6,5 @@ from .reset_agent_configuration import ResetAgentConfiguration
 from .propagation_credentials import PropagationCredentials
 from .ip_addresses import IPAddresses
 from .agent_configuration import AgentConfiguration
+from .pba_file_upload import PBAFileUpload, LINUX_PBA_TYPE, WINDOWS_PBA_TYPE
+from .pba_file_download import PBAFileDownload